+++ src/cheat/features/ServerLiar.h
@@ -30,7 +30,34 @@
 #define SERVERLIAR_H
 #pragma once
 
-extern BaseVariable cvarfilter_fps_max;
+extern VarBoolean cvarfilter_enable;
+extern VarBoolean cvarfilter_monitor_server;
+
+extern VarKeyValue cvarfilter_fps_max;
+extern VarKeyValue cvarfilter_fps_override;
+
+extern VarKeyValue fps_max;
+extern VarKeyValue fps_override;
+extern VarKeyValue gl_vsync;
+extern VarKeyValue developer;
+
+// movement
+extern VarKeyValue cl_movespeedkey;
+extern VarKeyValue cl_forwardspeed;
+extern VarKeyValue cl_sidespeed;
+extern VarKeyValue cl_upspeed;
+extern VarKeyValue cl_yawspeed;
+extern VarKeyValue cl_pitchspeed;
+
+// prediction
+extern VarKeyValue cl_lc;
+extern VarKeyValue cl_lw;
+
+// other
+extern VarKeyValue cl_minmodels;
+
+// networking
+extern VarKeyValue cl_cmdrate;
 
 class CServerLiar
 {
@@ -40,6 +67,26 @@ public:
 public:
 	// returns fake cvar value
 	const char* filter_cvarvalue(const char* cvar_name);
+
+	void render_ui();
+
+private:
+	void initialize();
+
+	void render_cvarfilter_slot(VarKeyValue* var);
+
+	// ik that this seems really dumb and out of place however, I don't really know how to properly
+	// work with input text buffers... xD
+	inline static constexpr size_t k_max_filtered_cvars = 32;
+	std::unordered_map<std::string, size_t> m_filtered_cvars;
+	char m_input_buffers[k_max_filtered_cvars][256];
+
+	bool keep_status_alive() const { return std::chrono::duration<float, std::ratio<1, 1>>(std::chrono::high_resolution_clock::now() - m_status_update).count() < 5; }
+
+	std::chrono::high_resolution_clock::time_point m_status_update;
+	std::string m_current_status;
+
+	std::vector<VarKeyValue*> m_variables;
 };
 
 #endif // SERVERLIAR_H